requirements.txt
Django==4.1
djangorestframework==3.13.1
pandas==1.4.3
requests==2.28.1
SQLAlchemy==1.4.40
xmltodict==0.13.0
